removed redundancies, moved methods to API controller and API routes
