+++ routes/api.php
@@ -3,6 +3,7 @@
 use App\Http\Controllers\Api;
 // use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\Licenses\LicenseCheckinController;
 
 
 /*
@@ -423,6 +424,12 @@ Route::group(['prefix' => 'v1', 'middleware' => ['api', 'throttle:'.config('app.
                 'licenses'
             ]
         )->name('api.assets.licenselist');
+
+
+      Route::get(
+          'api/v1/licenses/checkin-all/{license_id}',
+          [LicenseCheckinController::class, 'checkinLicense']
+      )->name('licenses/checkin-all');
 
         Route::get('bytag/{tag}',
             [
